*: rework to use ast generation vs templating code out
I felt like playing with go's ast package a bit, and here's what we
ended up with.

This version mostly works. It works for the example in the readme at
least, but doesn't yet work for variadic interfaces, or a few other edge
cases (like the primary interface being wrapped being defined in the
same package).